+++ romancal/outlier_detection/outlier_detection_step.py
@@ -1,7 +1,7 @@
 """Public common step definition for OutlierDetection processing."""
 
-import os
 from functools import partial
+from pathlib import Path
 
 from romancal.datamodels import ModelContainer
 from romancal.outlier_detection import outlier_detection
@@ -52,16 +52,20 @@ class OutlierDetectionStep(RomanStep):
     def process(self, input_models):
         """Perform outlier detection processing on input data."""
 
-        self.input_models = input_models
-        self.input_container = isinstance(self.input_models, ModelContainer)
         self.skip = False
 
-        # validation
-        if self.input_container and (
-            len(self.input_models) >= 2
-            and all(
-                model.meta.exposure.type == "WFI_IMAGE" for model in self.input_models
+        try:
+            self.input_models = ModelContainer(input_models)
+        except TypeError:
+            self.log.warning(
+                "Skipping outlier_detection - input cannot be parsed into a ModelContainer."
             )
+            self.skip = True
+            return input_models
+
+        # validation
+        if len(self.input_models) >= 2 and all(
+            model.meta.exposure.type == "WFI_IMAGE" for model in self.input_models
         ):
             # Setup output path naming if associations are involved.
             asn_id = self.input_models.asn_table.get("asn_id", None)
@@ -112,32 +116,33 @@ class OutlierDetectionStep(RomanStep):
             for model in self.input_models:
                 model.meta.cal_step["outlier_detection"] = state
                 if not self.save_intermediate_results:
-                    #  Remove unwanted files
-                    crf_path = self.make_output_path(basepath=model.meta.filename)
-                    # These lines to be used when/if outlier_i2d files follow
-                    # output_dir crf_file = os.path.basename(crf_path)
-                    # outlr_path = crf_path.replace(crf_file, outlr_file)
-                    outlr_file = model.meta.filename.replace("cal", "outlier_i2d")
-                    blot_path = crf_path.replace("crf", "blot")
-                    median_path = blot_path.replace("blot", "median")
+                    #  remove intermediate files found in
+                    #  make_output_path() and the local dir
+                    intermediate_files_paths = [
+                        Path(self.make_output_path()).parent,
+                        Path().cwd(),
+                    ]
+                    intermediate_files_suffixes = (
+                        "*blot.asdf",
+                        "*median.asdf",
+                        f'*{pars.get("resample_suffix")}*.asdf',
+                    )
+                    for current_path in intermediate_files_paths:
+                        for suffix in intermediate_files_suffixes:
+                            for filename in current_path.glob(suffix):
+                                filename.unlink()
+                                self.log.debug(f"    {filename}")
 
-                    for fle in [outlr_file, blot_path, median_path]:
-                        if os.path.isfile(fle):
-                            os.remove(fle)
-                            self.log.debug(f"    {fle}")
-
-            return self.input_models
-
-        # if input is not valid then log a warning message and
-        # skip outlier detection step
-        self.log.warning(
-            "Input is not a ModelContainer, \
-            does not contains >= 2 elements, \
-            or the elements contain the wrong exposure type \
-            (i.e. meta.exposure.type != 'WFI_IMAGE')."
-        )
-        if self.input_container:
+        else:
+            # if input can be parsed into a ModelContainer
+            # but is not valid then log a warning message and
+            # skip outlier detection step
+            self.log.warning(
+                "Skipping outlier_detection - at least two imaging observations are needed."
+            )
+            # set meta.cal_step.outlier_detection to SKIPPED
             for model in self.input_models:
                 model.meta.cal_step["outlier_detection"] = "SKIPPED"
-        self.skip = True
+            self.skip = True
+
         return self.input_models