+++ romancal/regtest/test_wfi_dq_init.py
@@ -68,13 +68,13 @@ def test_dq_init_image_step(rtdata, ignore_asdf_paths):
     assert diff.identical, diff.report()
 
 
-@metrics_logger("DMS26")
+@metrics_logger("DMS25")
 @pytest.mark.bigdata
 def test_dq_init_grism_step(rtdata, ignore_asdf_paths):
-    """DMS26 Test: Testing retrieval of best ref file for grism data,
+    """DMS25 Test: Testing retrieval of best ref file for grism data,
     and creation of a ramp file with CRDS selected mask file applied."""
 
-    input_file = "r0000201001001001002_01101_0001_WFI01_uncal.asdf"
+    input_file = "r0000201001001001001_01101_0001_WFI01_uncal.asdf"
     rtdata.get_data(f"WFI/grism/{input_file}")
     rtdata.input = input_file
 
@@ -82,34 +82,34 @@ def test_dq_init_grism_step(rtdata, ignore_asdf_paths):
     step = DQInitStep()
     model = rdm.open(rtdata.input)
     step.log.info(
-        "DMS26 MSG: Testing retrieval of best "
+        "DMS25 MSG: Testing retrieval of best "
         "ref file for grism data, "
         "Success is creation of a ramp file with CRDS selected "
         "mask file applied."
     )
 
-    step.log.info(f'DMS26 MSG: First data file: {rtdata.input.rsplit("/", 1)[1]}')
+    step.log.info(f'DMS25 MSG: First data file: {rtdata.input.rsplit("/", 1)[1]}')
     ref_file_path = step.get_reference_file(model, "mask")
     step.log.info(
-        f'DMS26 MSG: CRDS matched mask file: {ref_file_path.rsplit("/", 1)[1]}'
+        f'DMS25 MSG: CRDS matched mask file: {ref_file_path.rsplit("/", 1)[1]}'
     )
     ref_file_name = os.path.split(ref_file_path)[-1]
 
     assert "roman_wfi_mask" in ref_file_name
 
     # Test DQInitStep
-    output = "r0000201001001001002_01101_0001_WFI01_dqinit.asdf"
+    output = "r0000201001001001001_01101_0001_WFI01_dqinit.asdf"
     rtdata.output = output
     args = ["romancal.step.DQInitStep", rtdata.input]
     step.log.info(
-        "DMS26 MSG: Running data quality initialization step."
+        "DMS25 MSG: Running data quality initialization step."
         "The first ERROR is expected, due to extra CRDS parameters "
         "not having been implemented yet."
     )
     RomanStep.from_cmdline(args)
     ramp_out = rdm.open(rtdata.output)
     step.log.info(
-        "DMS26 MSG: Does ramp data contain pixeldq from mask file? :"
+        "DMS25 MSG: Does ramp data contain pixeldq from mask file? :"
         f' {("roman.pixeldq" in ramp_out.to_flat_dict())}'
     )
     assert "roman.pixeldq" in ramp_out.to_flat_dict()
@@ -117,7 +117,7 @@ def test_dq_init_grism_step(rtdata, ignore_asdf_paths):
     rtdata.get_truth(f"truth/WFI/grism/{output}")
     diff = compare_asdf(rtdata.output, rtdata.truth, **ignore_asdf_paths)
     step.log.info(
-        "DMS26 MSG: Was proper data quality initialized "
+        "DMS25 MSG: Was proper data quality initialized "
         "ramp data produced? : "
         f"{diff.identical}"
     )